Allow SYNTHETIC method when actual class is non-public.
This fixes a bug report "Example of run-process with redirection to pipe?"
(Kawa mailing list).  It turned out to involve synthetic methods,
but be somewhat tricky to (hopefully) fix.

* ClassType.java (isAccessible): Make static.
Interpret caller==null as being "artibitrary caller".

* ClassMethods.java (getMethod): Only override declared type
of method if implementationType is public.
(removeRedundantMethods): Remove SYNTHETIC if redundant.
It is not always redundant - consider StringBuilder.setLength,
which is a syntahic wrapper for AbstractStringBuilder.setLength.